[BZ 1009945] do not run callback on driver I/O thread and decrease throttling
The previous throttling defaults actually put too much throttling in place to
the point where it led to timeouts. There are two writes for raw data - one for
the actual value and one for the index update. The index update is done in a
future callback. The callback was running in one of the driver's I/O threads.
Because of the increased throttling, driver threads were getting tied up and
eventually timing out. Moving the callback to a separate thread pool resolves
the issue. I am also increasing the default throttling since it was way too
low.