Bug 1051413 - RFE: give plugin authors a tool to monitor MBeans of an application deployed on AS7 or EAP6
Added ApplicationMBeansComponent and ApplicationMBeansComponent. A separate project is needed because of classloading issues (the discovery class loader needs access to both the AS7 and JMX plugin classes).

The application MBeans container resource will be a service resource with a StandaloneAS or Managed AS parent.

Added a sample service MBean to javaee6-test-app.
Added mbeans-on-as7-plugin sample to demonstrate usage of the new facility