[BZ 1045589] make sure index entries get deleted
This commit fixes a couple issues in Aggregator. First, if waiting for the
index entries arrival was aborted, we were skipping the call to
deleteIndexEntries() which means one of the counters checked to determine if
aggregation is done wasn't getting updated. On a semi-related node, we might
be able to altogether get rid of the remainingIndexEntries counter. Secondly,
if the wait was aborted, we never obtained the index entries write lock but
were still making the calls to release it. This would result in an
IllegalMonitorStateException. Both of these issues have been resolved.

There is one other minor, significant change. The callback in
MetricsServer.updateMetricsIndex was running in an I/O thread. It is now
running in the aggregationWorkers thread pool.