BZ 1052390 - Clean up remoting wrapper to avoid race conditions if possible
Some unused or rarely constructors and methods were dropped.

The biggest change is in the client caching. The cache code is guaranteed to
call disconnect when a client is 'thrown away'. There is still a possibility
disconnect can happen in the middle of an invoke.

Original Author: Elias Ross <elias_ross@apple.com>
Signed-off-by: Jay Shaughnessy <jshaughn@redhat.com>
Applying this patch as-is, I see no issues with it and it cleans some
things up.  Relevant tests are passing.