[BZ 1045589] major refactoring of aggregation to limit concurrent reads
Even with really high throttling like only allowing 5k requests/sec, it was
still easily possible to do enough concurrent reads during aggregation that
could result in OMMs. Prior to this commit, several hundred or thousdand
queries could be basically executed in parallel to fetch raw data. The max
size of the payload coming in the form of a StorageResultSetFuture object is
about 30 KB.

With this commit, the number of concurrent reads that is done during
aggregation has a fix, configurable upper bound. It can be configured with the
following system properties,

    * rhq.metrics.aggregation.batch-size
    * rhq.metrics.aggregation.parallelism

Data for measurement schedules is fetched and computed in chunks or batches.
The batch-size property specifies the number of schedules for which data will
be fetched. The parallelism property specifies the number of batches that can
be processed in parallel. In terms of implementation, this is all handled using
a Semaphore. The number of semaphore permits is determined by
batch-size * parallelism. So for example, if batch-size is 25 and parallelism
is 4, then this will yield 100 permits, meaning at most 100 concurrent metric
data queries are allowed.