BZ-1049009 - JBoss Remoting SSL transport fails due to stale connections
This is a bug in JBoss Remoting 2.2.2.SP8 and was identified in
https://issues.jboss.org/browse/JBREM-1245. The fix from JBREM-1245 can not
easily be ported to 2.2.2 due to the code basis being different.

To fix this without needing a new version of JBoss Remoting, the rawSend
method has been updated to retry the request in the event that we receive
an RMI exception wrapped around the SSLException. As this is only done on
connection failure, the impact should be very low to normal operation of
the comm layer.
(cherry picked from commit 7160d4a1468f567654bea4ca60a625c3362a4492)

(cherry picked from commit 533a9ad74225fe952a1a553bf1c3588d98f35c06)