Bug 1029221 - Duplicate EAP resources are discovered if JBOSS_HOME can also be a symbolic link
The BZ was filed for EAP5 but the same problem occurs with EAP6

Use canonical path of configDir for resource key in EAP5
Use canonical path of hostConfigFile for resource key in EAP6
Old resource keys will be upgraded to preserve existing inventories

JBossInstanceInfo: removed usage of deprecated methods and made code easier to read.

*Note on testing EAP5*
The BZ report does not mention it but you won't be able to reproduce the problem if agent runs as root or as the same user as the managed server (this is because of the org.rhq.plugins.jbossas5.helper.JBossInstanceInfo.getHomeDir implementation).