[BZ 1045589] reduce defaults for aggregation batch size and parallelism
After more testing with a single storage node and doing as much as 50,000
metrics/minute, reducing the default batch size to 10 and parallelism to 4
has made things much more stable. I have not seen a single time out in the
several days since I have made the change, even when there are long running
compactions on the order of minutes.