need the row count for the index queries up front
We need the row count for the index queries up front before scheduling
aggregation tasks so that the remaining schedules variables can be set
and avoiding a potential race condition. If the count is not set up front, it
is possible for a BatchAggregator to finish and wind up setting the remaining
schedules variable to zero even when there are more schedules to process.