[BZ 1064110] make metrics aggregation settings configurable
Aggregation settings can be configured in two ways. The first is by directly in
editing the properties in rhq-server.properties. Doing so requires a server
restart for changes to take effect. The second way is through the Measurement
Subsystem resource which is defined in the rhq-server agent plugin. When
the settings are modified through the resource configuration, changes are
applied dynamically at runtime and do not require a server restart.