Bug 1059882 - Cluster-connections, bridges and JMS connection factories must define either connectors or discovery-group but not both
From 6.1 on, EAP has a guard which refuses modifications on these resources which include a value for both attributes, even if one of them is null. AS7 plugin had some code in ConnectorDiscoveryGroupValidatorComponent to check that both values were not filled, but allowed sending nulls or empty String.

Now in ConnectorDiscoveryGroupValidatorComponent configuration update happen in two phases:
1. execute a composite operation to set one of the two attributes and unset the other one
2. use the standard ConfigurationWriteDelegate for all other attributes

Side work: update ResourcesStandaloneServerTest and ResourcesDomainServerTest to exclude some resource types which are broken in some versions of EAP.