[1073201] Agent does a discovery loop if the inventory contains an unknown resource It's possible to disable a plugin on an agent after resources of that plugin's types are already in server-side inventory.  The server will report those resources to the agent during an inventory sync.  The agent will treat them as "unknown" resources because those resources will not have containers.
This fixes the handling for unknown resources with disabled resource types.
It ensures they are not merged into agent-side inventory and also do not
trigger further discovery scans.  Additionally, it now generates a
ResourceError for the server-side resource to help notify the user that the
resource is no longer being managed, and should be uninventoried.  Only
uninventory will stop the inventory sync overhead.