Bug 1082747 - Postgres Plugin - An Invalid Server Configuration Status Should Not Be Stack Trace
Throwing an Exception is the only way a plugin component can fail the process of starting of a resource. So without changing the agent plugin facet we can't rely avoid the stack trace shown on the server.

Still, the stack trace shown when a Postgres server resource is configured with invalid credentials was particularly long, so I changed the plugin component implentation to make the server show a stack trace with no cause and a condensed message ("Invalid password").