[BZ 1088264] AvailabilityExecutor stops calling getAvailability() on ResourceComponent after it previously failed with exception
The issue was in handling exception comming from future. When availability
check failed with exception we cought it, next run, just by calling
future.get() raises the very same exception. We forgot to mark future to be
rescheduled next time = setting it to null. This commit also makes exception
message more verbose so we know more what happened in plugin