add support for "decactivating" metrics_cache
Deactivating metrics_cache results in data *not* being pulled from it. This
applies to both past and current data. Once a specified time has been reached,
it is reactived and we start pulling from metrics_cache again. This logic is
necessary for two purrposes. First, we need it simplify and speed up the data
migration that will be necessary when upgrading to RHQ 4.11. Secondly, it is
necessary for the resizing of cache partitions to take effect. For example,
suppose we store data for up to 5 schedules per cache partition. We decide
to store data for 10 schedules per partition. We need to deactivate the cache
while the change takes effect.