[BZ 1015334] EAP 6 host controller domain host update results in managed servers no longer being "manageable"
Fixed javadocs, renamed variables. Also made access to ASHostName
synchronized since it's value can be changed from avail thread and is being
read all over other threads. Big change of this commit is getAvailability().
Now we directly read server's name attribute instead of 'launch-type' (which
was probably lets-just-read-something) and in case of failure (making we
just failed to read value, but connection and other stuf is OK) we re'read
<host name from file and try the request again. This approach is more robust
than previous, which was only able to detect the change in case server avail
went from DOWN to UP. Now, we can detect the change anytime.