[1070326] Do not individually track applications with the same WEB Context This is an initial commit for consideration.  It enhances discovery in the as7 plugin such that Deployments automatically process versioned artifact names.  For example, myapp-1.0.war results in a Deployment resource of myapp.way with version 1.0.  A subsequent deployment of myapp-2.0.war results in the existing myapp.war resource being upped to version 2.0.
This is applied to any Deployment based on the SubsystemDiscovery class.

Moreover, general support has been added such that when discovering a new
version of an existing resource, not only is the version updated, but the
resource's plugin configuration is updated with any properties set by the
discovery. In this way if the new version requires updated connection props
(like the EAP address path used to connect) it gets updated.  Since only
those props set by discovery are updated, any existing customizations will be
maintained in the existing configuration.