[BZ 1070277] - (JON3-39) JON should (be able to) block login to a user without roles
Adding a flag into system settings that will indicate if it is allowed
to log in without any assigned RHQ role. Our RBAC current implementation luckily protect user from
performing any server-side calls, but she is still able to see alert and
metric templates, download and install agent, etc.

If the flag (Enable Login Without Roles) is set to true everything is as before
this change. Otherwise, the user is not let into the application.
Because of the security reasons there is no difference between invalid
credentials and valid credentials together with  no assigned role.

It defaults to false.

NOTE: It does not apply only for LDAP users but for all users, because
it should be consistent for both JDBC and LDAP users.