[BZ 1069547] Modifying the Bundle creation workflow to allow for no files provided by the bundle recipe.
Also, getting the list of bundle file names will not throw exception
when bundle parsing fails - it will merely return empty list in that
case.

The first change is there to support the bundle types that cannot
figure out the list of the files required in the bundle from the
recipe alone.

The second change is for bundle types that cannot even properly parse
the recipe in isolation from the files in the distribution file.