make sure we drop the legacy metrics_index table during schema update
The metrics_index table wasn't getting dropped for new installs. Fixing this
created a problem in TopologyManager where it sets gc_grace_seconds for each of
the tables. This commit modifies the logic in TopologyManager as well so that
it no longer tries to update gc_grace_seconds for metrics_index and so that it
does set it for metrics_cache and metrics_cache_index.