BZ1094424 - do not reindex tables by default; don't use UNRECOVERABLE
ALTER INDEX {0} REBUILD UNRECOVERABLE

When being run on Oracle, this prevents inserts while the rebuild takes place,
which may be several minutes. This has caused agents to time out and false
alerts to be generated.

Unrecoverable is not a good idea as it locks the table during the calculation:
https://community.oracle.com/thread/1103546?tstart=0

The default should be false, as well UNRECOVERABLE is not a good idea either on
Oracle because of the locking.