refactoring cache index querying and making raw data age limit configurable
I have introduced paging when querying the index for the current time slice.
The query is against a single partition which can get large and cause read
timeouts. Paging should help guard against that.

I have also made the age limit for raw data that is ingested configurable.
This means for example, that the user can decide whether or not the server
should accept or reject measurement reports with data that is already 5 days
old.