[BZ 1078437] Correct the navigation in content subsystem JSF pages
The /portal prefix is only valid in the <redirect/> navigation cases
because in those cases it is represented literally as the URL to go to.

In all other cases it is a logical name that wasn't affected by the move of
the JSF UI to the /portal context root.

(cherry picked from commit d3fb7a117a2f11cb16256eea9cb6ed7ded222049)
Signed-off-by: Lukas Krejci <lkrejci@redhat.com>