[BZ 1071330] Try to avoid timeouts on retrieving system settings
This is a bit of a shot in the dark to reduce the possibility of
timeouts when reading system settings in the server under load.

The concurrency on the bean has been changed from the default
container managed with a write lock on each method to bean managed
with a single critical section in the cache loading and writing methods.

Also, the deobfuscation of the password fields happens only on the
cache reload and only if the settings were changed.

All this to minimize the contention on the SystemManagerBean and to
speed up the retrieval of the settings as much as possible.

(cherry picked from commit 2d896974f527317fe418b88fcab59c3e19d1230d)
Signed-off-by: Libor Zoubek <lzoubek@redhat.com>

Conflicts:
	modules/core/domain/src/main/java/org/rhq/core/domain/common/composite/SystemSetting.java