JBossAS7: export "Domain Name" as a trait
We need this in order to have dynagroup definitions composed from Domain
Name + some other configuration property. Also, this property is called
'name' on HC, but in RHQ we call it 'domain-name' - this is because
dynagroup evaluator is not strict when it comes to trait names - so having
resource.trait[name] includes several traits.