[BZ 873471] Token replacement takes into account <property> properties
The ant users are used to defining properties using the standard ant
<property> task. We now recognize such properties when replacing
tokens in the bundle files.

To support <property file="..." /> format, the file in question has
to be defined as a bundle file using rhq:file or rhq:url-file. This
would make it deployed though, which might not be what the bundle
author wants, so a new attribute "deploy" has been added to all the
rhq:file, rhq:archive, rhq:url-file and rhq:url-archive, defaulting
to true, meaning whether to deploy the file/archive or not.

Also, the standard Ant <property> task can now appear inside
rhq:bundle alongside the rhq:input-property.

An example deploy.xml might look like this:

<?xml version="1.0"?>
<project name="property-demo" default="main"
         xmlns:rhq="antlib:org.rhq.bundle">

    <rhq:bundle name="Property loading demo" version="1.0">
      <rhq:input-property name="user.provided"/>
      <property name="bundle.defined" value="xxx"/>
      <property file="in-bundle.properties"/>

      <rhq:deployment-unit name="sample-deployment-unit"
                           compliance="full">
        <rhq:file name="in-bundle.properties" deploy="false"/>
        <rhq:file name="deployed.file" replace="true" />
      </rhq:deployment-unit>
    </rhq:bundle>
    <target name="main"/>
</project>

The deployed.file mentioned above could contain the
@@bundle.defined@@ token which would be replaced with "xxx". It could
also contain tokens with names of any properties defined in the
"in-bundle.properties" file that is supposed to be placed in the
bundle file alongside the deploy.xml.