[BZ 1069547] Removing support for patches from the content subsystem
This is to not overload the user with overlapping yet incompatible
choices for doing the same thing.

After discussion, we decided to keep the support for patching in the
bundle subsystem and drop it from the content subsystem.

This also means we need to drop the specialized rollback operation
because using that we could bring RHQ bundles into an inconsistent
state.