[1119262 jon3-7] EAP6 drifts exclude pattern not working for windows case The issue here is that the drift template defines the plugin descriptor and in this case, the associated excludes filter: pattern="*_history/**/*". It can only be predefined in one format, and the plugin naturally uses linux-style forward slashes for the directory separator.
But on Windows that is the wrong separator.

The solution is to convert to the native separator slash internally at
"detection-time", regardless of the slashing used in the pattern.  This is,
I think, in-line with ant file-pattern approach we are using.

Also,
-- added new FileUtil.useNativeSlash(String path)

Cherry-Pick of master 83e926c0debb558b2e90efee47317517ea4bb4db