[BZ 1120418] add config association to maintenance step
Adding Configuration association to MaintenanceStep to allow for storing
arbitrary data with each step. There are at least a couple things we need to
store - resource operation arguments and a snapshot of the cluster. By snapshot
I mean a mapping of all of the nodes. This will be necessary to determine
whether or not steps need to be recalculated.