[BZ 1120418] set job params and some support for calculating steps
This is the initial commit for StorageMaintenanceJob. It is is a container for
job steps that provides some abstractions for effecting changes on a job as a
whole.

So far I am only working with node deployment. When a new node is imported into
inventory and linked to a resource, StorageNodeManagerBean schedules a
maintenance job. It now passes params to the job via a Configuration object.

Steps are calculated using a StepCalulator. DeployCaclulcator is the first
impl. It only creates the announce step so far. The step calculator is also
responsible for persisting and/or deleting steps, which is why I thought it
makes more sense to make them EJBs.