[BZ 1120418] update deploy steps calculator
This is a first cut at end to end functionality for the deployment steps
calculator. It now generates tests for updating C* schema and running cluster
maintenance. There is plenty of refactoring to be done, but I first wanted to
see how one (or more) of the step calculators would look.