[BZ 1120418] initial commit for RepairService
This is the first in a series of changes to improve how we run repair. Adding
Adding a JMX notification listener to track failed repair sessions so that we
can subsequently retry repair only on those failed ranges. This is using
straight JMX because I could not get notifications working with EMS. I am not
sure that I want to use EMS though because of
https://bugzilla.redhat.com/show_bug.cgi?id=960773.