[BZ 1120418] abort failed announce job when announce operation fails
The failed step job includes running announce followed by repair. If announce
fails, then it most likely means that the node and/or agent is down, in which
case repair will fail as well. Normally the announce step has a failure
strategy of continue, but in the case of the failed announce job we want to
abort.

I have changed StepRunner to support this use case. The exceute() method no
longer takes a MaintenanceStep argument. There is now a setter for the step.
This makes the step and its configuration more easily available to the
getFailureStrategy() method.