[BZ 1120418] move creation of new job for failed step into the step runner
Prior to this commit when a step whose failure strategy is CONTINUE fails, we
called the StepCalculator to produce a new job for the failed step. The
StepRunner now produces the new job. There are a couple of reasons for the
change. First, it eliminates branching logic that would be necessary in the
StepCalculator. Secondly, the step runner may have info from the failed
execution that needs to be included as input to the new job. This is the case
with repair. When a repair operation fails, we want to get the failed ranges
from the operation history results, and pass those as params to a new job to
rerun repair on said ranges. That operation history is already available to
the step runner. It would have been difficult to make it availble to the
StepCalculator.