rhqctl install was failing with issues like:   "Could not complete storage cluster schema installation: Authentication error on host"
The issue was intermittent.  The problem seems to be that the latest
version of Cassandra is taking longer than earlier versions to
initialize its default files (and the default cassandra user).
This happens at first startup.  We start cassandra just prior
to installing the server, because the server install applies the
schema updates.  But it seems cassandra was not quite ready at
times.  This change simply introduces a short delay to allow
cassandra time to finish its first-startup-initialization.

Additionally, an unrelated fix to ensure the yaml file is
properly closed on config updates.