Bug 999494 - Import child Resource: jboss as (both eap 4 and 5) are not being imported to inventory list after successful import operation
Not only tell the user a request has been submitted, but if the request failed or succeed and why.

So now the user knows if nothing happens because the resource in already in the inventory, in the discovery queue, or ignored.

(cherry picked from commit 748df2aff13be656a7d1677d43eebb535f782e27)
Signed-off-by: Thomas Segismont <tsegismo@redhat.com>